Fix deleting wrong keys from admin panel
- Fixed wrong_keys.html template using WrongKeys.team instead of
  WrongKeys.teamid
- Added the id of the wrong key to the wrong_keys.html template entries
- Changed /admin/wrong_keys to only take the key id instead of team and
  challenge id, to allow deleting the intended key when a team has more
  than one wrong key for a challenge. (Previously it would delete only
  the first wrong key entry for that challenge)

+++ CTFd/admin.py
@@ -705,14 +705,17 @@ def delete_solve(teamid, chalid):
     return '1'
 
 
-@admin.route('/admin/wrong_keys/<teamid>/<chalid>/delete', methods=['POST'])
+@admin.route('/admin/wrong_keys/<keyid>/delete', methods=['POST'])
 @admins_only
-def delete_wrong_key(teamid, chalid):
-    wrong_key = WrongKeys.query.filter_by(teamid=teamid, chalid=chalid).first()
+def delete_wrong_key(keyid):
+    wrong_key = WrongKeys.query.filter_by(id=keyid).first()
+    if (wrong_key == None):
+        return '0'
     db.session.delete(wrong_key)
     db.session.commit()
     db.session.close()
     return '1'
+
 
 
 @admin.route('/admin/statistics', methods=['GET'])
@@ -755,7 +758,7 @@ def admin_wrong_key(page='1'):
     page_start = results_per_page * ( page - 1 )
     page_end = results_per_page * ( page - 1 ) + results_per_page
 
-    wrong_keys = WrongKeys.query.add_columns(WrongKeys.chalid, WrongKeys.flag, WrongKeys.teamid, WrongKeys.date,\
+    wrong_keys = WrongKeys.query.add_columns(WrongKeys.id, WrongKeys.chalid, WrongKeys.flag, WrongKeys.teamid, WrongKeys.date,\
                 Challenges.name.label('chal_name'), Teams.name.label('team_name')).\
                 join(Challenges).join(Teams).order_by('team_name ASC').slice(page_start, page_end).all()
